libkernel: Cleanup some function places
