+++ src/core/libraries/kernel/event_queue.h
@@ -7,10 +7,12 @@
 #include <mutex>
 #include <string>
 #include <vector>
-
 #include <boost/asio/steady_timer.hpp>
-
 #include "common/types.h"
+
+namespace Core::Loader {
+class SymbolsResolver;
+}
 
 namespace Libraries::Kernel {
