+++ src/core/libraries/kernel/time_management.cpp
@@ -4,7 +4,6 @@
 #include <thread>
 
 #include "common/assert.h"
-#include "common/debug.h"
 #include "common/native_clock.h"
 #include "core/libraries/error_codes.h"
 #include "core/libraries/kernel/time_management.h"
@@ -258,7 +257,33 @@ Common::NativeClock* GetClock() {
 
 } // namespace Dev
 
-void timeSymbolsRegister(Core::Loader::SymbolsResolver* sym) {
+int PS4_SYSV_ABI sceKernelConvertUtcToLocaltime(time_t time, time_t* local_time,
+                                                struct OrbisTimesec* st, unsigned long* dst_sec) {
+    LOG_TRACE(Kernel, "Called");
+#ifdef __APPLE__
+    // std::chrono::current_zone() not available yet.
+    const auto* time_zone = date::current_zone();
+#else
+    const auto* time_zone = std::chrono::current_zone();
+#endif
+    auto info = time_zone->get_info(std::chrono::system_clock::now());
+
+    *local_time = info.offset.count() + info.save.count() * 60 + time;
+
+    if (st != nullptr) {
+        st->t = time;
+        st->west_sec = info.offset.count() * 60;
+        st->dst_sec = info.save.count() * 60;
+    }
+
+    if (dst_sec != nullptr) {
+        *dst_sec = info.save.count() * 60;
+    }
+
+    return ORBIS_OK;
+}
+
+void RegisterTime(Core::Loader::SymbolsResolver* sym) {
     clock = std::make_unique<Common::NativeClock>();
     initial_ptc = clock->GetUptime();
     LIB_FUNCTION("4J2sUJmuHZQ", "libkernel", 1, "libkernel", 1, 1, sceKernelGetProcessTime);
@@ -284,6 +309,7 @@ void timeSymbolsRegister(Core::Loader::SymbolsResolver* sym) {
     LIB_FUNCTION("lLMT9vJAck0", "libScePosix", 1, "libkernel", 1, 1, posix_clock_gettime);
     LIB_FUNCTION("smIj7eqzZE8", "libScePosix", 1, "libkernel", 1, 1, posix_clock_getres);
     LIB_FUNCTION("0NTHN1NKONI", "libkernel", 1, "libkernel", 1, 1, sceKernelConvertLocaltimeToUtc);
+    LIB_FUNCTION("-o5uEDpN+oY", "libkernel", 1, "libkernel", 1, 1, sceKernelConvertUtcToLocaltime);
 }
 
 } // namespace Libraries::Kernel
